remoe console logs and add search param utils
